+++ CHANGELOG.md
@@ -8,9 +8,11 @@ Unreleased
 ************
 * Added caching to improve validate SBDefinition performance
 * Add `GET /prjs/{prj_id}/status` and `GET /sbds/{sbd_id}/status`
-* Add the `prj_status` to the return value of the `/odt/prsls/project-view` API
-* [BREAKING] Delete the unused `/sbds/validate` API in favour of the validation service `/validate/sbds`
 * Adds PUT /prjs/{prj_id}/status/ready and PUT /prjs/{prj_id}/status/draft
+* Add the `prj_status` to the return value of the `/odt/prsls/project-view` API
+* Changes to `PUT /pht/panels/{panel_id}` to enforce proposal–panel exclusivity and correct proposal status transitions.
+* [BREAKING] Delete the unused `/sbds/validate` API in favour of the validation service `/validate/sbds`
+* [BUGFIX] Return CORS headers on errors
 
 12.1.0
 ************